Fixes #34.

+++ api/src/main/scala/org/platanios/tensorflow/api/io/FileIO.scala
@@ -477,7 +477,8 @@ object FileIO {
     try {
       rename(temporaryFilePath, filePath, overwrite)
     } finally {
-      deleteFile(temporaryFilePath)
+      if (exists(temporaryFilePath))
+        deleteFile(temporaryFilePath)
     }
   }
 }
